% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shooting_fu.R, R/shooting_scd.R
\name{euclidnorm}
\alias{euclidnorm}
\alias{euclidnorm}
\title{Computes the euclidean norm}
\usage{
euclidnorm(x)

euclidnorm(x)
}
\arguments{
\item{x}{a number or vector}

\item{x}{a number or vector}
}
\value{
the euclidean norm

the euclidean norm
}
\description{
Computes the euclidean norm

Computes the euclidean norm
}
\examples{
euclidnorm(3)
euclidnorm(3:4)
euclidnorm(c(3,4))
euclidnorm(3)
euclidnorm(3:4)
euclidnorm(c(3,4))
}
