% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shooting_scd.R
\name{shooting_scd}
\alias{shooting_scd}
\title{Shooting SCD algorithm}
\usage{
shooting_scd(X, y, lambda, eps = 1e-06, max_steps = 1000)
}
\arguments{
\item{X}{design matrix}

\item{y}{vector of responses}

\item{lambda}{tuning parameter}

\item{eps}{tolerance}
}
\value{
The vector of parameters, steps performed and indicator of convergence
}
\description{
Shooting SCD algorithm
}
